changes: add Select and Command BP native events; add function GetHit…

+++ Source/GBTeam6/Game/GameStateDefault.cpp
@@ -39,7 +39,10 @@ void AGameStateDefault::InitializeServices() {
 
 void AGameStateDefault::ClearServices() {
 	UE_LOG(LgService, Log, TEXT("<%s>: Clearing Services"), *GetNameSafe(this));
-	this->MappingService->DestroyService();
+	if(this->MappingService->IsValidLowLevelFast())
+	{
+		this->MappingService->DestroyService();
+	}
 	this->MappingService = nullptr;
 }
